CodeAssist/backend/requirements.txt
flask
flask_cors
docker
psycopg2-binary
python-dotenv
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy